Move MappingAnalytics to proper package
