+++ legend-engine-xt-analytics-mapping-api/src/main/java/org/finos/legend/engine/api/analytics/model/MappedEntityInfo.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.finos.legend.engine.api.model;
+package org.finos.legend.engine.api.analytics.model;
 
 import java.util.List;
