+++ legend-engine-xt-analytics-mapping-api/src/main/java/org/finos/legend/engine/api/analytics/model/MappingModelCoverageAnalysisResult.java
@@ -12,19 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.finos.legend.engine.api.model;
+package org.finos.legend.engine.api.analytics.model;
 
 import java.util.List;
 
-public class MappedEntity
+public class MappingModelCoverageAnalysisResult
 {
-    public String path;
-    public List<MappedProperty> properties;
-    public MappedEntityInfo info;
+    public List<MappedEntity> mappedEntities;
 
-    public MappedEntity(String path, List<MappedProperty> properties)
+    public MappingModelCoverageAnalysisResult()
     {
-        this.path = path;
-        this.properties = properties;
+        // DO NOT DELETE: this resets the default constructor for Jackson
+    }
+
+    public MappingModelCoverageAnalysisResult(List<MappedEntity> mappedEntities)
+    {
+        this.mappedEntities = mappedEntities;
     }
 }
